fix(ci): line-limit check — split provider readiness tests
